Main Activity broken to fragment
