+++ app/src/main/java/com/ranjithnaidu/android/playbook/utils/Extensions.kt
@@ -1,7 +1,18 @@
 package com.ranjithnaidu.android.playbook.utils
 
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.MutableLiveData
 
 fun <T> MutableLiveData<T>.setValueIfNew(newValue: T) {
     if (this.value != newValue) value = newValue
+}
+
+/**
+ * Allows calls like
+ *
+ * `supportFragmentManager.inTransaction { add(...) }`
+ */
+inline fun FragmentManager.inTransaction(func: FragmentTransaction.() -> FragmentTransaction) {
+    beginTransaction().func().commit()
 }
