--- app/src/main/res/layout/activity_main.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-                name="viewModel"
-                type="com.ranjithnaidu.android.playbook.post.viewmodel.PostsViewModel"/>
-
-    </data>
-    <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            tools:context=".home.view.MainActivity">
-
-        <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/posts_recycler_view"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"/>
-
-    </FrameLayout>
-</layout>
