+++ app/src/main/res/layout/fragment_posts_list.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <data>
+
+        <variable
+                name="viewModel"
+                type="com.ranjithnaidu.android.playbook.post.viewmodel.PostsViewModel"/>
+
+    </data>
+
+    <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/posts_recycler_view"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"/>
+
+</layout>
